+++ lib/vagrant/command/status.rb
@@ -1,17 +1,16 @@
 module Vagrant
   module Command
-    class StatusCommand < Base
-      argument :name, :type => :string, :optional => true
+    class StatusCommand < NamedBase
       register "status", "Shows the status of the current Vagrant environment."
 
       def route
         state = nil
-        results = env.vms.collect do |name, vm|
+        results = target_vms.collect do |vm|
           state ||= vm.created? ? vm.vm.state.to_s : "not_created"
-          "#{name.to_s.ljust(25)}#{state.gsub("_", " ")}"
+          "#{vm.name.to_s.ljust(25)}#{state.gsub("_", " ")}"
         end
 
-        state = env.vms.length == 1 ? state : "listing"
+        state = target_vms.length == 1 ? state : "listing"
 
         env.ui.info(I18n.t("vagrant.commands.status.output",
                     :states => results.join("\n"),